flight_assistant_pro/requirements.txt
streamlit
pandas
duckdb
google-generativeai
faiss-cpu
python-dotenv
networkx